Removed dependency on rwk_parse.h
- Two new functions added: nn_nchar() and nn_str2array(). The latter
  is identical to that originally used in rwk_parse.h. The former is
  similar to rwk_countcols but is generalized to all cases and returns
  the number of chars found (rather than the number of columns: nchar + 1).

- nn_char() returns an unsigned long, so as much as possible has been
  updated to reflect this. There are however functions called from liblar
  that may be affected.

- nn_limits.h added so that maximum buffer size, maximum number of rows etc...
  can be tracked and changed accordingly.

+++ include/nn_objects.h
@@ -36,7 +36,7 @@ extern void create_network(struct NeuralNetwork *nnet, int nlayers, int *nnodes)
 extern void free_network(struct NeuralNetwork *nnet);
 
 struct TrainingData {
-  int nobs;
+  unsigned long nobs;
   int ninputs;
   int noutputs;
   struct lar_matrix inputs;
